Move lock unlock wipe mdm tests to separate file
